requirements.txt
aniso8601==10.0.1
anyio==4.9.0
apispec==6.8.2
appdirs==1.4.4
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
Flask-RESTful==0.3.10
flask-restx==1.3.0
flask-smorest==0.46.1
git-changelog==2.5.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
logging==0.4.9.6
MarkupSafe==3.0.2
marshmallow==4.0.0
notion-client==2.3.0
packaging==25.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
semver==3.0.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.4.0
webargs==8.7.0
Werkzeug==3.1.3
aniso8601==10.0.1
anyio==4.9.0
apispec==6.8.2
appdirs==1.4.4
attrs==25.3.0
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
flake8==7.2.0
Flask==3.1.1
Flask-RESTful==0.3.10
flask-restx==1.3.0
flask-smorest==0.46.1
git-changelog==2.5.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
logging==0.4.9.6
MarkupSafe==3.0.2
marshmallow==4.0.0
mccabe==0.7.0
mypy_extensions==1.1.0
notion-client==2.3.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.13.0
pyflakes==3.3.2
Pygments==2.19.1
pytest==8.4.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
semver==3.0.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.4.0
webargs==8.7.0
Werkzeug==3.1.3
